feat: implement wallet details sidebar
- Added sidebar animation for wallet details and transaction history.
- Created separate TransactionList and TransactionDetails components for better organization.
- Moved TransactionHistoryModal and TransactionList to a new transaction directory.
- Updated imports to reflect new file structure.
- Removed the old TransactionHistoryModal implementation.
- Introduced a new type definition for Transaction in a dedicated types file.
- Enhanced wallet details UI with copy address functionality and improved layout.
- Updated the animated components for smoother transitions.
