+++ app/components/WalletDetails.tsx
@@ -1,36 +1,46 @@
 "use client";
-import { useRef, useState } from "react";
-import { AnimatePresence, motion } from "framer-motion";
-
-import { useOutsideClick } from "../hooks";
-import {
-  classNames,
-  fetchSupportedTokens,
-  formatCurrency,
-  detectWalletProvider,
-} from "../utils";
+import { useState } from "react";
+import { classNames, formatCurrency, shortenAddress } from "../utils";
 import { useBalance } from "../context/BalanceContext";
-import { dropdownVariants } from "./AnimatedComponents";
 import { usePrivy } from "@privy-io/react-auth";
-import { Token } from "../types";
 import { useNetwork } from "../context/NetworksContext";
 import { TransferModal } from "./TransferModal";
-import { ArrowDown01Icon, Wallet01Icon } from "hugeicons-react";
+import {
+  ArrowRight03Icon,
+  ArrowUp04Icon,
+  Copy01Icon,
+  Wallet01Icon,
+  ArrowLeft02Icon,
+  ArrowDown01Icon,
+} from "hugeicons-react";
 import Image from "next/image";
 import { FundWalletModal } from "./FundWalletModal";
 import { useFundWalletHandler } from "../hooks/useFundWalletHandler";
 import { useInjectedWallet } from "../context";
+import { toast } from "sonner";
+import { Dialog } from "@headlessui/react";
+import { AnimatePresence, motion } from "framer-motion";
+import { sidebarAnimation } from "./AnimatedComponents";
+import { TransactionList } from "./transaction/TransactionList";
+import { TransactionDetails } from "./transaction/TransactionDetails";
+import type { Transaction } from "./transaction/types";
+import { PiCheck } from "react-icons/pi";
 
 export const WalletDetails = () => {
-  const [isOpen, setIsOpen] = useState<boolean>(false);
   const [isTransferModalOpen, setIsTransferModalOpen] =
     useState<boolean>(false);
   const [isFundModalOpen, setIsFundModalOpen] = useState(false);
+  const [activeTab, setActiveTab] = useState<"balances" | "transactions">(
+    "balances",
+  );
+  const [isSidebarOpen, setIsSidebarOpen] = useState(false);
+  const [selectedTransaction, setSelectedTransaction] =
+    useState<Transaction | null>(null);
+  const [isAddressCopied, setIsAddressCopied] = useState(false);
 
   const { selectedNetwork } = useNetwork();
   const { allBalances } = useBalance();
   const { isInjectedWallet, injectedAddress } = useInjectedWallet();
-
   const { user } = usePrivy();
 
   const { handleFundWallet } = useFundWalletHandler("Wallet details");
@@ -56,152 +66,245 @@ export const WalletDetails = () => {
     );
   };
 
-  const dropdownRef = useRef<HTMLDivElement>(null);
-  useOutsideClick({
-    ref: dropdownRef,
-    handler: () => setIsOpen(false),
-  });
+  const handleSidebarClose = () => {
+    setIsSidebarOpen(false);
+    setSelectedTransaction(null);
+  };
 
-  const tokens: { name: string; imageUrl: string | undefined }[] = [];
-  const fetchedTokens: Token[] =
-    fetchSupportedTokens(selectedNetwork.chain.name) || [];
-  for (const token of fetchedTokens) {
-    tokens.push({
-      name: token.symbol,
-      imageUrl: token.imageUrl,
-    });
-  }
+  const handleCopyAddress = () => {
+    navigator.clipboard.writeText(activeWallet?.address ?? "");
+    setIsAddressCopied(true);
+    toast.success("Address copied to clipboard");
+    setTimeout(() => setIsAddressCopied(false), 2000);
+  };
 
-  const getTokenImageUrl = (tokenName: string) => {
-    const token = tokens.find((token) => token.name === tokenName);
-    return token ? token.imageUrl : "";
+  const handleBackToList = () => {
+    setSelectedTransaction(null);
   };
 
   return (
     <>
-      <div ref={dropdownRef} className="relative">
-        <button
-          type="button"
-          title="Wallet balance"
-          onClick={() => {
-            setIsOpen(!isOpen);
-          }}
-          className="flex h-9 items-center justify-center gap-2 rounded-xl bg-accent-gray px-2.5 py-2.5 transition-colors duration-300 hover:bg-border-light focus:outline-none focus-visible:ring-2 focus-visible:ring-lavender-500 focus-visible:ring-offset-2 focus-visible:ring-offset-white dark:bg-white/10 dark:hover:bg-white/20 dark:focus-visible:ring-offset-neutral-900 sm:py-0"
-        >
-          <Wallet01Icon className="size-5 text-icon-outline-secondary dark:text-white/50" />
-          <div className="h-9 w-px border-r border-dashed border-border-light dark:border-white/10" />
-          <div className="flex items-center gap-1.5 dark:text-white/80">
-            <p>{formatCurrency(activeBalance?.total ?? 0, "USD", "en-US")}</p>
-            <ArrowDown01Icon
-              aria-label="Caret down"
-              className={classNames(
-                "mx-1 size-4 text-icon-outline-secondary transition-transform duration-300 dark:text-white/50",
-                isOpen ? "rotate-180" : "",
-              )}
-            />
-          </div>
-        </button>
+      <button
+        type="button"
+        title="Wallet balance"
+        onClick={() => {
+          setIsSidebarOpen(!isSidebarOpen);
+        }}
+        className="flex h-9 items-center justify-center gap-2 rounded-xl bg-accent-gray px-2.5 py-2.5 transition-colors duration-300 hover:bg-border-light focus:outline-none focus-visible:ring-2 focus-visible:ring-lavender-500 focus-visible:ring-offset-2 focus-visible:ring-offset-white dark:bg-white/10 dark:hover:bg-white/20 dark:focus-visible:ring-offset-neutral-900 sm:py-0"
+      >
+        <Wallet01Icon className="size-5 text-icon-outline-secondary dark:text-white/50" />
+        <div className="h-9 w-px border-r border-dashed border-border-light dark:border-white/10" />
+        <div className="flex items-center gap-1.5 dark:text-white/80">
+          <p>{formatCurrency(activeBalance?.total ?? 0, "USD", "en-US")}</p>
+          <ArrowDown01Icon
+            aria-label="Caret down"
+            className={classNames(
+              "mx-1 size-4 text-icon-outline-secondary transition-transform duration-300 dark:text-white/50",
+              isSidebarOpen ? "rotate-180" : "",
+            )}
+          />
+        </div>
+      </button>
 
-        <AnimatePresence>
-          {isOpen && (
-            <motion.div
-              initial="closed"
-              animate="open"
-              exit="closed"
-              variants={dropdownVariants}
-              className="absolute right-0 mt-3 w-[273px] space-y-2 rounded-xl border border-neutral-100 bg-white p-2 shadow-lg dark:border-white/5 dark:bg-surface-overlay"
-            >
-              <AnimatePresence>
-                {activeBalance?.balances ? (
-                  <motion.div
-                    initial="closed"
-                    animate="open"
-                    exit="closed"
-                    variants={dropdownVariants}
-                    className="space-y-3 rounded-xl bg-accent-gray p-3 dark:bg-white/5"
-                  >
+      <AnimatePresence>
+        {isSidebarOpen && (
+          <Dialog
+            as="div"
+            className="fixed inset-0 z-50 overflow-hidden"
+            onClose={handleSidebarClose}
+            open={isSidebarOpen}
+          >
+            <div className="flex h-full">
+              <motion.div
+                initial={{ opacity: 0 }}
+                animate={{ opacity: 1 }}
+                exit={{ opacity: 0 }}
+                className="fixed inset-0 bg-black/30 backdrop-blur-sm"
+                onClick={handleSidebarClose}
+              />
+
+              <motion.div
+                {...sidebarAnimation}
+                className="z-50 my-4 ml-auto mr-4 flex h-[calc(100%-32px)] w-full max-w-[396px] flex-col overflow-hidden rounded-[20px] border border-border-light bg-white shadow-lg dark:border-white/5 dark:bg-surface-overlay"
+              >
+                {selectedTransaction ? (
+                  <div className="flex h-full flex-col p-6">
+                    <div className="mb-6 flex items-center gap-3">
+                      <button
+                        type="button"
+                        title="Back to transactions"
+                        onClick={handleBackToList}
+                        className="flex items-center gap-2 text-sm font-medium text-text-body dark:text-white"
+                      >
+                        <ArrowLeft02Icon className="size-5 text-outline-gray dark:text-white/50" />
+                        Back
+                      </button>
+                    </div>
+                    <div className="scrollbar-hide -mx-6 flex-1 overflow-y-auto px-5">
+                      <TransactionDetails transaction={selectedTransaction} />
+                    </div>
+                  </div>
+                ) : (
+                  <div className="flex h-full flex-col p-5">
                     <div className="flex items-center justify-between">
-                      <h3 className="font-light text-gray-500 dark:text-white/50">
-                        {isInjectedWallet
-                          ? detectWalletProvider()
-                          : "Noblocks Wallet"}
-                      </h3>
+                      <h2 className="text-lg font-semibold text-text-body dark:text-white">
+                        Wallet
+                      </h2>
+                      <button
+                        type="button"
+                        title="Close wallet details"
+                        onClick={handleSidebarClose}
+                        className="rounded-lg p-2 transition-colors hover:bg-accent-gray dark:hover:bg-white/10"
+                      >
+                        <ArrowRight03Icon className="size-5 text-outline-gray dark:text-white/50" />
+                      </button>
                     </div>
 
-                    <ul className="space-y-2 text-neutral-900 dark:text-white/80">
-                      {Object.entries(activeBalance?.balances || {}).map(
-                        ([token, balance]) => (
-                          <li key={token} className="flex items-center gap-1">
-                            {(() => {
-                              const imageUrl = getTokenImageUrl(token);
-
-                              return imageUrl ? (
-                                <Image
-                                  src={imageUrl}
-                                  alt={token}
-                                  width={14}
-                                  height={14}
-                                  className="size-3.5"
-                                />
-                              ) : null;
-                            })()}
-
-                            <span className="font-medium">
-                              {balance} {token}
-                            </span>
-                          </li>
-                        ),
-                      )}
-                    </ul>
-
-                    {!isInjectedWallet && (
-                      <div className="flex items-center gap-3">
+                    <div className="mt-6 space-y-6 rounded-[20px] border border-border-light bg-transparent p-4 dark:border-white/10">
+                      <div className="flex items-center justify-between gap-2">
+                        <div className="flex items-center gap-2">
+                          <Wallet01Icon className="size-4 text-outline-gray dark:text-white/50" />
+                          <p className="text-text-body dark:text-white/80">
+                            {shortenAddress(activeWallet?.address ?? "", 8)}
+                          </p>
+                        </div>
                         <button
-                          type="button"
-                          onClick={() => {
-                            setIsFundModalOpen(true);
-                            setIsOpen(false);
-                          }}
-                          className="font-medium text-lavender-500"
+                          onClick={handleCopyAddress}
+                          title="Copy wallet address"
+                          className="rounded-lg p-2 transition-colors hover:bg-accent-gray dark:hover:bg-white/10"
                         >
-                          Fund
-                        </button>
-                        <p className="text-[10px] text-gray-200 dark:text-white/10">
-                          |
-                        </p>
-                        <button
-                          type="button"
-                          onClick={() => {
-                            setIsTransferModalOpen(true);
-                            setIsOpen(false);
-                          }}
-                          className="font-medium text-lavender-500"
-                        >
-                          Transfer
+                          {isAddressCopied ? (
+                            <PiCheck className="size-4 text-green-500" />
+                          ) : (
+                            <Copy01Icon className="size-4 text-outline-gray dark:text-white/50" />
+                          )}
                         </button>
                       </div>
-                    )}
-                  </motion.div>
-                ) : (
-                  <motion.div
-                    initial="closed"
-                    animate="open"
-                    exit="closed"
-                    variants={dropdownVariants}
-                    className="space-y-3 rounded-xl bg-accent-gray p-3 dark:bg-white/5"
-                  >
-                    <div className="flex items-center justify-between">
-                      <h3 className="font-light text-gray-500 dark:text-white/50">
-                        No balances available
-                      </h3>
+
+                      <div className="text-2xl font-medium text-text-body dark:text-white">
+                        {formatCurrency(
+                          activeBalance?.total ?? 0,
+                          "USD",
+                          "en-US",
+                        )}
+                      </div>
+
+                      {!isInjectedWallet && (
+                        <div className="grid grid-cols-2 gap-4">
+                          <button
+                            type="button"
+                            title="Transfer funds"
+                            onClick={() => setIsTransferModalOpen(true)}
+                            className="min-h-11 w-full rounded-xl bg-accent-gray py-2 text-sm font-medium text-gray-900 transition-all hover:scale-[0.98] hover:bg-[#EBEBEF] active:scale-95 dark:bg-white/5 dark:text-white dark:hover:bg-white/10"
+                          >
+                            Transfer
+                          </button>
+                          <button
+                            type="button"
+                            title="Fund wallet"
+                            onClick={() => setIsFundModalOpen(true)}
+                            className="min-h-11 w-full rounded-xl bg-accent-gray py-2 text-sm font-medium text-gray-900 transition-all hover:scale-[0.98] hover:bg-[#EBEBEF] active:scale-95 dark:bg-white/5 dark:text-white dark:hover:bg-white/10"
+                          >
+                            Fund
+                          </button>
+                        </div>
+                      )}
                     </div>
-                  </motion.div>
+
+                    <div className="mt-6 flex items-center gap-6">
+                      <button
+                        type="button"
+                        onClick={() => setActiveTab("balances")}
+                        title="View balances"
+                        className={classNames(
+                          "text-sm font-medium",
+                          activeTab === "balances"
+                            ? "text-text-body dark:text-white"
+                            : "text-text-disabled dark:text-white/30",
+                        )}
+                      >
+                        Balances
+                      </button>
+                      <button
+                        type="button"
+                        onClick={() => setActiveTab("transactions")}
+                        title="View transactions"
+                        className={classNames(
+                          "text-sm font-medium",
+                          activeTab === "transactions"
+                            ? "text-text-body dark:text-white"
+                            : "text-text-disabled dark:text-white/30",
+                        )}
+                      >
+                        Transactions
+                      </button>
+                    </div>
+
+                    <div className="mt-6 flex-grow overflow-y-scroll">
+                      {activeTab === "balances" && activeBalance?.balances && (
+                        <div className="h-full space-y-4 overflow-y-auto pb-16">
+                          {Object.entries(activeBalance.balances).map(
+                            ([token, balance]) => (
+                              <div
+                                key={token}
+                                className="flex items-center justify-between text-sm"
+                              >
+                                <div className="flex items-center gap-3">
+                                  <div className="relative">
+                                    <Image
+                                      src={`/logos/${token.toLowerCase()}-logo.svg`}
+                                      alt={token}
+                                      width={32}
+                                      height={32}
+                                      className="size-8 rounded-full"
+                                    />
+                                    <Image
+                                      src={selectedNetwork.imageUrl}
+                                      alt={selectedNetwork.chain.name}
+                                      width={16}
+                                      height={16}
+                                      className="absolute -bottom-1 -right-1 size-4 rounded-full"
+                                    />
+                                  </div>
+                                  <div className="flex flex-col">
+                                    <span className="text-text-body dark:text-white/80">
+                                      {token}
+                                    </span>
+                                    <span className="text-text-secondary dark:text-white/50">
+                                      {balance}
+                                    </span>
+                                  </div>
+                                </div>
+                                <div className="flex flex-col items-end">
+                                  <span className="text-text-body dark:text-white/80">
+                                    $1,234.56
+                                  </span>
+                                  <div className="flex items-center gap-1">
+                                    <ArrowUp04Icon className="size-3 text-green-500" />
+                                    <span className="text-sm text-text-secondary dark:text-white/50">
+                                      2.5%
+                                    </span>
+                                  </div>
+                                </div>
+                              </div>
+                            ),
+                          )}
+                        </div>
+                      )}
+
+                      {activeTab === "transactions" && (
+                        <TransactionList
+                          onSelectTransaction={setSelectedTransaction}
+                        />
+                      )}
+                    </div>
+                  </div>
                 )}
-              </AnimatePresence>
-            </motion.div>
-          )}
-        </AnimatePresence>
-      </div>
+              </motion.div>
+            </div>
+          </Dialog>
+        )}
+      </AnimatePresence>
 
       {!isInjectedWallet && (
         <>
