+++ app/types.ts
@@ -246,3 +246,22 @@ export type Network = {
         dark: string;
       };
 };
+
+export interface Transaction {
+  id: string;
+  type: string;
+  amount: string;
+  currency: string;
+  swappedCurrency?: string;
+  nativeValue: string;
+  time: string;
+  status: string;
+  fees?: string;
+  day?: string;
+  recipient?: string;
+  bank?: string;
+  account?: string;
+  memo?: string;
+  fundStatus?: string;
+  timeSpent?: string;
+}
